requirements.txt
pytest==8.3.2
requests==2.32.3
selenium==4.23.1
allur==2.30.0
allure-pytest==2.13.5
